# Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beacon Nonprofit Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /in-kind-donation-items:
    get:
      operationId: listInKindDonationItems
      tags: [donations]
      summary: List in-kind donation items
      parameters:
        - name: donationId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      security:
        - bearerAuth: []
      responses:
        "200":
          description: In-kind donation items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InKindDonationItem"
    post:
      operationId: createInKindDonationItem
      tags: [donations]
      summary: Create in-kind donation item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInKindDonationItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InKindDonationItem"
  /in-kind-donation-items/{id}:
    get:
      operationId: getInKindDonationItem
      tags: [donations]
      summary: Get in-kind donation item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InKindDonationItem"
    delete:
      operationId: deleteInKindDonationItem
      tags: [donations]
      summary: Delete in-kind donation item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /impact-snapshots:
    get:
      operationId: listImpactSnapshots
      tags: [impactSnapshots]
      summary: List impact snapshots
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: published
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Impact snapshot list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSnapshotListResponse"
    post:
      operationId: createImpactSnapshot
      tags: [impactSnapshots]
      summary: Create impact snapshot (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateImpactSnapshotBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSnapshot"
  /impact-snapshots/{id}:
    get:
      operationId: getImpactSnapshot
      tags: [impactSnapshots]
      summary: Get impact snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Impact snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSnapshot"
    patch:
      operationId: updateImpactSnapshot
      tags: [impactSnapshots]
      summary: Update impact snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImpactSnapshotBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSnapshot"
    delete:
      operationId: deleteImpactSnapshot
      tags: [impactSnapshots]
      summary: Delete impact snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /impact-snapshots/{id}/publish:
    post:
      operationId: publishImpactSnapshot
      tags: [impactSnapshots]
      summary: Publish impact snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSnapshot"
  /impact-snapshots/{id}/unpublish:
    post:
      operationId: unpublishImpactSnapshot
      tags: [impactSnapshots]
      summary: Unpublish impact snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpactSnapshot"

  # ─── ML PREDICTIONS ───────────────────────────────────────────────────────
  /ml/predictions:
    get:
      operationId: listMlPredictions
      tags: [mlPredictions]
      summary: List ML prediction snapshots
      parameters:
        - name: pipeline
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
        - name: entityType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: ML predictions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MlPredictionListResponse"
...
  /ml/predictions/{entityType}/{entityId}:
    get:
      operationId: getEntityPredictions
      tags: [mlPredictions]
      summary: Get all predictions for a specific entity
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entity predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MlPrediction"

  # ─── REPORTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ───────────────────────────────────────────────────────────────
...
    ImpactSnapshot:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        period:
          type: string
        isPublished:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        residentsServed:
          type: integer
        totalDonationsAmount:
          type: number
        programOutcomes:
          type: object
          additionalProperties: true
        safehousetsCovered:
          type: integer
        reintegrationCount:
          type: integer
        summary:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, period, isPublished, residentsServed, totalDonationsAmount, safehousetsCovered, reintegrationCount]
    CreateImpactSnapshotBody:
      type: object
      properties:
        title:
          type: string
        period:
          type: string
        residentsServed:
          type: integer
        totalDonationsAmount:
          type: number
        programOutcomes:
          type: object
          additionalProperties: true
        safehousetsCovered:
          type: integer
        reintegrationCount:
          type: integer
        summary:
          type: string
      required: [title, period, residentsServed, totalDonationsAmount, safehousetsCovered, reintegrationCount]
    UpdateImpactSnapshotBody:
      type: object
      properties:
        title:
          type: string
        period:
          type: string
        residentsServed:
          type: integer
        totalDonationsAmount:
          type: number
        programOutcomes:
          type: object
          additionalProperties: true
        safehousetsCovered:
          type: integer
        reintegrationCount:
          type: integer
        summary:
          type: string
    ImpactSnapshotListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ImpactSnapshot"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── ML ──────────────────────────────────────────────────────────────────
    MlPrediction:
      type: object
      properties:
        id:
          type: integer
        pipeline:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        predictionValue:
          type: number
        confidenceScore:
          type: number
        riskBand:
          type: ["string", "null"]
        topFeatures:
          type: array
          items:
            type: object
            properties:
              feature:
                type: string
              importance:
                type: number
            required: [feature, importance]
        recommendedAction:
          type: ["string", "null"]
        modelVersion:
          type: string
        predictedAt:
          type: string
      required: [id, pipeline, entityType, entityId, predictionValue, confidenceScore, modelVersion, predictedAt]
    MlPredictionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MlPrediction"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
...
    InKindDonationItem:
      type: object
      properties:
        id:
          type: integer
        donationId:
          type: integer
        itemDescription:
          type: string
        category:
          type: string
          enum: [food, clothing, medicine, equipment, educational, hygiene, other]
        quantity:
          type: integer
        unit:
          type: string
          nullable: true
        estimatedValuePerUnit:
          type: number
          nullable: true
        totalEstimatedValue:
          type: number
          nullable: true
        condition:
          type: string
          enum: [new, good, fair, poor]
        receivedBy:
          type: integer
          nullable: true
        receivedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, donationId, itemDescription, category, quantity, condition, createdAt, updatedAt]
    CreateInKindDonationItemBody:
      type: object
      properties:
        donationId:
          type: integer
        itemDescription:
          type: string
        category:
          type: string
          enum: [food, clothing, medicine, equipment, educational, hygiene, other]
        quantity:
          type: integer
        unit:
          type: string
        estimatedValuePerUnit:
          type: number
        totalEstimatedValue:
          type: number
        condition:
          type: string
          enum: [new, good, fair, poor]
        receivedBy:
          type: integer
        receivedAt:
          type: string
          format: date-time
        notes:
          type: string
      required: [donationId, itemDescription, category, quantity, condition]